--- core/experimental/SwipeGesture.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: fd84445d8a604ec39a2623f01659119e
-timeCreated: 1509140705
