--- core/experimental/UpdatedControls.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: e771d78fd67289841b369853a38fb5e4
-timeCreated: 1509138412
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
