+++ core/input/Desktop/MoveCameraWithMouse.cs.meta
