+++ core/input/Tools/CreateObjectTool.cs
@@ -5,9 +5,8 @@ using WorldWizards.core.controller.level;
 using WorldWizards.core.controller.level.utils;
 using WorldWizards.core.entity.coordinate.utils;
 using WorldWizards.core.entity.gameObject;
-using WorldWizards.core.experimental.controllers;
 
-namespace worldWizards.core.experimental.controllers.Tools
+namespace worldWizards.core.input.Tools
 {
     public class CreateObjectTool : Tool
     {
