+++ core/input/Tools/CreateObjectTool.cs.meta
