+++ core/input/Tools/StandardTool.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using UnityEngine;
-using WorldWizards.core.experimental.controllers;
 
-namespace worldWizards.core.experimental.controllers.Tools
+namespace worldWizards.core.input.Tools
 {
     public class StandardTool : Tool
     {
