+++ core/input/Tools/StandardTool.cs.meta
